Fixed the memory leak!!1
The memory leak was due to our misuse of the NHibernate sessionfactory object. We kept spwaning new session factories for every request.
The memory used by all the sessionfactories was never released (by some unknown reference holder) and hence the large mem-usage.
The fix in this commit is to share a single(ton) SessionFactory between all web services. This is implemented in the DefaultSessionFactory class.