Huge commit:
Refactored the HolidayProvider into its own domain model with a separate repository available for widgets
Removed two projects and merged the HolidayXmlParser into the integration project
Refactored the Iteration domain model to not use a holidayprovider, but take a list of holidays instead
Refactored the workingdaysleft controller to use the newly available IRepository<Holiday> to fetch holidays

NOTE: The holiday repository does not contain weekends! It is now the clients responsibility to filer out weekends if they need to.