Upgrading WiX and fixing correct .NET version for App pool on IIS7 install.
By upgrading to newest version of WiX 3.5 we enable the use of the
ManagedRuntimeVersion property on IIS Application pools. This property
allows WiX to easily set the runtime version to be used for apps in the
given app pool.

Note: This change is ignored on older versions of IIS and is thus only a
perfect fix on IIS7. For older versions we need to run some custom rolled
tasks.